+++ cosigner/cosigner_test.go
@@ -1,0 +1,48 @@
+package cosigner
+
+import (
+	"testing"
+	"time"
+
+	"github.com/lestrrat-go/jwx/v2/jwa"
+	"github.com/openpubkey/openpubkey/pktoken"
+	"github.com/openpubkey/openpubkey/pktoken/mocks"
+	"github.com/openpubkey/openpubkey/util"
+)
+
+func TestSimpleCosigner(t *testing.T) {
+	// Generate the key pair for our cosigner
+	alg := jwa.ES256
+	signer, err := util.GenKeyPair(alg)
+	if err != nil {
+		t.Error(err)
+	}
+
+	cos := &Cosigner{
+		Alg:    alg,
+		Signer: signer,
+	}
+
+	email := "arthur.aardvark@example.com"
+	pkt, err := mocks.GenerateMockPKTokenWithEmail(signer, alg, email)
+
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	cosignerClaims := pktoken.CosignerClaims{
+		Iss:         "example.com",
+		KeyID:       "none",
+		Algorithm:   cos.Alg.String(),
+		AuthID:      "none",
+		AuthTime:    time.Now().Unix(),
+		IssuedAt:    time.Now().Unix(),
+		Expiration:  time.Now().Add(time.Hour).Unix(),
+		RedirectURI: "none",
+	}
+
+	err = cos.Cosign(pkt, cosignerClaims)
+	if err != nil {
+		t.Error(err)
+	}
+}
