+++ cosigner/msgs/msgs.go
@@ -1,0 +1,7 @@
+package msgs
+
+type InitMFAAuth struct {
+	RedirectUri string `json:"ruri"`
+	TimeSigned  int64  `json:"time"`
+	Nonce       string `json:"nonce"`
+}
